feature:OSIS-7 Delete S3 Credentials unit tests
Description:
1. Delete S3 Credentials unit tests
2. Refactor User tests and S3Credential tests
